requirements.txt
langchain==0.1.5
openai==1.11.0
huggingface_hub==0.20.3
transformers==4.37.2
sentence-transformers==2.3.1
uvicorn==0.27.0.post1
ctransformers==0.2.27
python-box==7.1.1
Jinja2==3.1.3
python-dotenv==1.0.1
streamlit==1.31.0
ipykernel==6.29.0  # used to set a notebook environment